requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
requests
beautifulsoup4
python-dateutil